refactor:関数名の変更・ファイル位置の変更
