+++ app/circles/[circle_id]/activities/[activity_id]/page.tsx
@@ -2,7 +2,7 @@ import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { getMembershipRequests } from "@/actions/circle/membership-request"
 import { auth } from "@/auth"
 import { CircleDetailPage } from "@/components/layouts/circle-detail-page"
-import { demo } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/components/tab/tab-title"
 import { getActivities, getActivityById } from "@/data/activity"
 
 interface Props {
@@ -13,7 +13,7 @@ interface Props {
 }
 
 export const generateMetadata = ({ params }: Props) =>
-  demo(params.circle_id || "", "活動日程")
+  MetadataSet(params.circle_id || "", "活動日程")
 
 /*export const generateMetadata = async ({ params }: Props) => {
   const { circle_id } = params
