+++ app/circles/[circle_id]/activities/new/page.tsx
@@ -2,14 +2,14 @@ import { Center } from "@yamada-ui/react"
 import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { auth } from "@/auth"
 import { ActivityForm } from "@/components/forms/activity-form"
-import { demo } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/components/tab/tab-title"
 
 interface Props {
   params: { circle_id?: string }
 }
 
 export const generateMetadata = ({ params }: Props) =>
-  demo(params.circle_id || "", "新規活動日程")
+  MetadataSet(params.circle_id || "", "新規活動日程")
 
 export const generateStaticParams = async () => {
   const circles = await getCircles()
