+++ app/circles/[circle_id]/announcement/[topic_id]/page.tsx
@@ -2,7 +2,7 @@ import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { getMembershipRequests } from "@/actions/circle/membership-request"
 import { auth } from "@/auth"
 import { CircleDetailPage } from "@/components/layouts/circle-detail-page"
-import { demo } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/components/tab/tab-title"
 import { getAnnouncementById, getAnnouncements } from "@/data/announcement"
 
 interface Props {
@@ -13,7 +13,7 @@ interface Props {
 }
 
 export const generateMetadata = ({ params }: Props) =>
-  demo(params.circle_id || "", "お知らせ")
+  MetadataSet(params.circle_id || "", "お知らせ")
 
 export const generateStaticParams = async () => {
   const circles = await getCircles()
