+++ app/circles/[circle_id]/thread/[topic_id]/page.tsx
@@ -2,7 +2,7 @@ import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { getMembershipRequests } from "@/actions/circle/membership-request"
 import { auth } from "@/auth"
 import { CircleDetailPage } from "@/components/layouts/circle-detail-page"
-import { demo } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/components/tab/tab-title"
 import { getThreadById, getThreads } from "@/data/thread"
 
 interface Props {
@@ -13,7 +13,7 @@ interface Props {
 }
 
 export const generateMetadata = ({ params }: Props) =>
-  demo(params.circle_id || "", "スレッド")
+  MetadataSet(params.circle_id || "", "スレッド")
 
 export const generateStaticParams = async () => {
   const circles = await getCircles()
