+++ app/circles/[circle_id]/thread/create/page.tsx
@@ -2,13 +2,13 @@ import { Center } from "@yamada-ui/react"
 import { getCircleById, getCircles } from "@/actions/circle/fetch-circle"
 import { auth } from "@/auth"
 import { ThreadForm } from "@/components/forms/thread-form"
-import { demo } from "@/components/tab/tab-title"
+import { MetadataSet } from "@/components/tab/tab-title"
 interface Props {
   params: { circle_id?: string }
 }
 
 export const generateMetadata = ({ params }: Props) =>
-  demo(params.circle_id || "", "新規スレッド作成")
+  MetadataSet(params.circle_id || "", "新規スレッド作成")
 
 export const generateStaticParams = async () => {
   const circles = await getCircles()
